// src/api/generated/model/nutrientDefinitionCategory.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Fitness API
 * REST API for the Fitness & Nutrition SPA. Call GET /sanctum/csrf-cookie then POST /api/auth/login before accessing protected endpoints.
 * OpenAPI spec version: 1.0.0
 */

export type NutrientDefinitionCategory = typeof NutrientDefinitionCategory[keyof typeof NutrientDefinitionCategory];


export const NutrientDefinitionCategory = {
  energy: 'energy',
  macro: 'macro',
  mineral: 'mineral',
  vitamin: 'vitamin',
  other: 'other',
} as const;