// src/api/generated/model/nutrientReferenceDirection.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Fitness API
 * REST API for the Fitness & Nutrition SPA. Call GET /sanctum/csrf-cookie then POST /api/auth/login before accessing protected endpoints.
 * OpenAPI spec version: 1.0.0
 */

export type NutrientReferenceDirection = typeof NutrientReferenceDirection[keyof typeof NutrientReferenceDirection];


export const NutrientReferenceDirection = {
  at_least: 'at_least',
  informational: 'informational',
  reduce_if_above: 'reduce_if_above',
} as const;