// src/api/generated/model/nutrientReferenceReferenceType.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Fitness API
 * REST API for the Fitness & Nutrition SPA. Call GET /sanctum/csrf-cookie then POST /api/auth/login before accessing protected endpoints.
 * OpenAPI spec version: 1.0.0
 */

export type NutrientReferenceReferenceType = typeof NutrientReferenceReferenceType[keyof typeof NutrientReferenceReferenceType];


export const NutrientReferenceReferenceType = {
  RDA: 'RDA',
  AI: 'AI',
  CDRR: 'CDRR',
} as const;