// src/api/generated/model/periodNutrientIntakeStatus.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Fitness API
 * REST API for the Fitness & Nutrition SPA. Call GET /sanctum/csrf-cookie then POST /api/auth/login before accessing protected endpoints.
 * OpenAPI spec version: 1.0.0
 */

export type PeriodNutrientIntakeStatus = typeof PeriodNutrientIntakeStatus[keyof typeof PeriodNutrientIntakeStatus];


export const PeriodNutrientIntakeStatus = {
  no_data: 'no_data',
  unknown: 'unknown',
  partial: 'partial',
  complete: 'complete',
} as const;